+++ test/integration/src/site/_collections/fixtures/test-5.md
@@ -9,4 +9,4 @@ web_lighthouse:
 # A test post
 
 This test post belongs to collections:
-+ recentBlogposts
++ recentBlogPosts
