+++ test/integration/src/site/_collections/recent-blog-posts/collection.md
@@ -1,0 +1,3 @@
+{%- for post in collections.recentBlogPosts -%}
+<p>{{ post.fileSlug }}</p>
+{%- endfor -%}
