+++ test/integration/src/site/_collections/recent-blog-posts/prod.js
@@ -11,9 +11,9 @@ const outputPath = path.join(
   'index.html',
 );
 
-describe('recent-blogposts', function() {
-  it('includes 3 most recent blog posts that have a hero image or a thumbnail, ordered by date', async function() {
-    const expected = '<p>test-5</p><p>test-4</p><p>test-3</p>';
+describe('recent-blog-posts', function() {
+  it('does not include drafts', async function() {
+    const expected = '<p>test-5</p><p>test-4</p><p>test-2</p>';
     const actual = fs.readFileSync(outputPath, 'utf8');
     assert.equal(actual, expected);
   });
