--- test/integration/src/site/_collections/recent-blogposts/collection.md
@@ -1,3 +1,0 @@
-{%- for post in collections.recentBlogposts -%}
-<p>{{ post.fileSlug }}</p>
-{%- endfor -%}
